Пересмотрел реализацию методов -int minValue(int[] values) -List<Integer> oddOrEven(List<Integer> integers)
